Servidor/Monitormanto de canais/requirements.txt
# Video Qualifier - Python Dependencies
# Liftlio Production
# Python 3.12+

# ============================================
# Core Framework
# ============================================
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1

# ============================================
# HTTP Client
# ============================================
httpx==0.27.2

# ============================================
# YouTube API (Official Google)
# ============================================
google-api-python-client==2.155.0
google-auth==2.38.0
google-auth-httplib2==0.2.0

# ============================================
# Claude AI (Anthropic)
# ============================================
anthropic==0.42.0

# ============================================
# Supabase
# ============================================
supabase==2.9.1

# ============================================
# Data Validation
# ============================================
pydantic==2.10.4
pydantic-settings==2.7.1

# ============================================
# Logging
# ============================================
loguru==0.7.3

# ============================================
# Testing
# ============================================
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
